refactor are-valid-ids util
