+++ controllers/fractie.ts
@@ -1,5 +1,5 @@
 import { fractie } from '../data-access/fractie';
-import { areIdsValid, RDF_TYPE } from '../util/are-ids-valid';
+import { areIdsValid, RDF_TYPE } from '../util/valid-id';
 import { STATUS_CODE } from '../util/constants';
 import { HttpError } from '../util/http-error';
